Chunky commit :(
Re-implemented project using yarn as opposed to npm. React templating now started to be implemented, now abstracted the different a-frame components as jsx elements. Globe now has object mapped to the surface after bug fix where the boat was not being mapped correctly to the globe due to a difference in radius mapping.